html/semantics/scripting-1/the-script-element/module/dynamic-import/microtasks/with-import-assertions.html

<!DOCTYPE html>
<title>Dynamic import interaction with microtask queue</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
import { ticker } from "./ticker.js";

promise_test(async t => {
  const getCount = ticker(1000);

  try {
    await import("./../../imports-a.js", { type: "<invalid>" });
    assert_unreached('import() should reject');
  } catch (_) {}

  assert_less_than(getCount(), 1000);
}, "import() should not drain the microtask queue if it fails while validating the 'type' assertion");
</script>
